[rend2] Get rid of most gcc warnings on rend2 compile

+++ codemp/rd-rend2/tr_init.cpp
@@ -490,7 +490,7 @@ static const char *GetGLExtensionsString()
 	}
 
 	*p = '\0';
-	assert((p - extensionString) == extensionStringLen);
+	assert((size_t)(p - extensionString) == extensionStringLen);
 
 	return extensionString;
 }
